Moved SetupDB to test code package
